requirements.txt
astroid==2.2.5
beautifulsoup4==4.7.1
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
google==2.0.2
google-api-core==1.8.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-cloud-vision==0.36.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpcio==1.19.0
httplib2==0.12.1
idna==2.8
isort==4.3.15
lazy-object-proxy==1.3.1
lxml==4.3.2
mccabe==0.6.1
numpy==1.16.2
opencv-python==4.0.0.21
Pillow==5.4.1
protobuf==3.7.0
public==2018.11.20
pyasn1==0.4.5
pyasn1-modules==0.2.4
pylint==2.3.1
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==1.8
temp==2018.11.20
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.2
wget==3.2
wrapt==1.11.1
